Change SegmentMetadata.getCrc and getDataCrc to return long
The CRC is a long everywhere it is stored: SegmentMetadataImpl reads both from
creation.meta with DataInputStream.readLong() and holds them in long fields,
SegmentZKMetadata exposes them as longs, and the segment metadata JSON emits
them as numbers. Only the SegmentMetadata accessors rendered that long as a
String, so most callers parsed it straight back.

Return long from both accessors and drop the parse at those call sites. The
same type is carried through the in-process APIs that pass a CRC around:
SegmentDirectoryLoaderContext, SegmentDataManager.getCrc,
BaseTableDataManager.initSegmentDirectory, ImmutableSegmentLoader.preprocessSegment,
the expected-CRC arguments of MinionTaskUtils.getValidDocIdFromServerMatchingCrc
and the expected-CRC list of UpsertCompactMergeTaskExecutor.

No serialization format changes. Every CRC that reaches a payload, a path or a
config value is rendered with Long.toString at that boundary, so the ValidDocIds
responses, the /segments/crc endpoint, the ZK payload and the minion task
configs are byte-for-byte unchanged.

SegmentCrcVirtualColumnProvider loses its null and NumberFormatException
branches, which are unreachable once the accessor is typed. A case in
UpsertCompactMergeTaskExecutorTest asserting that a null segment CRC is
rejected is dropped, since that input is no longer representable.

TablesResourceTest also switches to statically imported TestNG assertions,
which accounts for most of its diff.

+++ pinot-segment-local/src/test/java/org/apache/pinot/segment/local/segment/index/SegmentMetadataImplTest.java
@@ -108,7 +108,7 @@ public class SegmentMetadataImplTest {
 
     JsonNode jsonMeta = metadata.toJson(null);
     assertEquals(jsonMeta.get("segmentName").asText(), metadata.getName());
-    Assert.assertEquals(jsonMeta.get("crc").asLong(), Long.valueOf(metadata.getCrc()).longValue());
+    Assert.assertEquals(jsonMeta.get("crc").asLong(), metadata.getCrc());
     Assert.assertTrue(jsonMeta.get("creatorName").isNull());
     assertEquals(jsonMeta.get("creationTimeMillis").asLong(), metadata.getIndexCreationTime());
     assertEquals(jsonMeta.get("timeColumn").asText(), metadata.getTimeColumn());
